+++ .idea/runConfigurations/LersosaPythonApplication.xml
@@ -1,13 +1,13 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="LersosaPythonApplication" type="Python.FastAPI">
-    <option name="file" value="D:\CodesProject\Java\LeyramuLersosa\lersosa-python\app\main.py" />
+    <option name="file" value="D:\CodesProject\Java\LeyramuLersosa\lersosa-python\src\app\app.py" />
     <module name="lersosa-python" />
     <option name="ENV_FILES" value="" />
     <option name="INTERPRETER_OPTIONS" value="" />
     <option name="PARENT_ENVS" value="true" />
     <option name="SDK_HOME" value="$PROJECT_DIR$/.venv/Scripts/python.exe" />
     <option name="SDK_NAME" value="Python 3.13 (LeyramuLersosa)" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/lersosa-python" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/lersosa-python/src" />
     <option name="IS_MODULE_SDK" value="true" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
