+++ lersosa-python/src/app/__init__.py
@@ -4,5 +4,5 @@
 #  The author disclaims all warranties, express or implied, including but not limited to the warranties of merchantability and fitness for a particular purpose. Under no circumstances shall the author be liable for any special, incidental, indirect, or consequential damages arising from the use of this software.
 #  By using this project, users acknowledge and agree to abide by these terms and conditions.
 
-from .code_status import CodeStatus
-from .msg_status import MsgStatus
+
+from .app import app
