+++ lersosa-python/src/app/common/repo_result.py
