+++ lersosa-python/src/app/enum/__init__.py
@@ -4,4 +4,6 @@
 #  The author disclaims all warranties, express or implied, including but not limited to the warranties of merchantability and fitness for a particular purpose. Under no circumstances shall the author be liable for any special, incidental, indirect, or consequential damages arising from the use of this software.
 #  By using this project, users acknowledge and agree to abide by these terms and conditions.
 
-from .tset_service import TestService
+
+from .code_status import CodeStatus
+from .msg_status import MsgStatus
