+++ lersosa-python/src/app/service/__init__.py
@@ -4,24 +4,5 @@
 #  The author disclaims all warranties, express or implied, including but not limited to the warranties of merchantability and fitness for a particular purpose. Under no circumstances shall the author be liable for any special, incidental, indirect, or consequential damages arising from the use of this software.
 #  By using this project, users acknowledge and agree to abide by these terms and conditions.
 
-import inspect
-import os
-from importlib import import_module
 
-from app.base import BaseController
-
-current_dir = os.path.dirname(os.path.abspath(__file__))
-package_name = __package__
-
-controllers = BaseController()
-
-for file in os.listdir(current_dir):
-    if file.endswith('.py') and not file.startswith("__"):
-        module_name = file[:-3]
-        full_module_name = f"{package_name}.{module_name}" if package_name else module_name
-        imported_module = import_module(full_module_name)
-        for name, obj in inspect.getmembers(imported_module):
-            if inspect.isclass(obj) and issubclass(obj, BaseController):
-                controllers.include_router(obj())
-
-__all__ = ["controllers"]
+from .tset_service import TestService
