+++ lersosa-python/src/main.py
@@ -4,16 +4,10 @@
 #  The author disclaims all warranties, express or implied, including but not limited to the warranties of merchantability and fitness for a particular purpose. Under no circumstances shall the author be liable for any special, incidental, indirect, or consequential damages arising from the use of this software.
 #  By using this project, users acknowledge and agree to abide by these terms and conditions.
 
-import os
 
+import uvicorn
 
-# Nacos 配置类
-class NacosConfig:
-    NACOS_SERVER_ADDR = os.getenv('NACOS_SERVER_ADDR', 'localhost:8848')
-    NACOS_NAMESPACE = os.getenv('NACOS_NAMESPACE', '356d484c-399c-4a23-9419-e200e8edbff9')
-    NACOS_GROUP = os.getenv('NACOS_GROUP', 'SERVICE_GROUP')
-    NACOS_USERNAME = os.getenv('NACOS_USERNAME', 'nacos')
-    NACOS_PASSWORD = os.getenv('NACOS_PASSWORD', 'Zcx@223852//')
-    NACOS_DATA_ID = os.getenv('NACOS_DATA_ID', 'lersosa-service-python.yml')
-    NACOS_SERVICE_NAME = os.getenv('NACOS_SERVICE_NAME', 'lersosa-service-python')
-    HEARTBEAT_INTERVAL = os.getenv('HEARTBEAT_INTERVAL', 10)
+from app import app
+
+if __name__ == "__main__":
+    uvicorn.run(app, host="127.0.0.1", port=8001, log_level="info")
